+++ Sources/HarborUtils/Errors.swift
@@ -1,0 +1,8 @@
+public enum HarborError: Error {
+    case invalidDirectory
+    case invalidFile
+    case invalidPath
+    case fleetNotFound
+    case packageNotFound
+    case buildFailed
+}